fix anchor point
