+++ .codesandbox/templates/vanilla/src/testcases/polygons.ts
@@ -44,12 +44,13 @@ export function testCase(canvas: fabric.Canvas) {
     {
       top: 300,
       stroke: 'blue',
-      strokeWidth: 6,
+      strokeWidth: 40,
       scaleX: 2,
       scaleY: 1,
       strokeUniform: true,
       exactBoundingBox: true,
-      // skewX: 10,
+      strokeMiterLimit: 9999,
+      skewX: 10,
       transparentCorners: false,
       cornerStyle: 'circle',
       hasBorders: false,
